Fix mishandling of a final event with blank Text field
If the last subtitle Event had a blank Text field, and there was another
Section following, the parser would gobble up the two newlines as part
of the Event separator, and then refuse to parse the next section header
"[Section]" because it's expecting another event to follow.

This commit replaces instance where it was expecting generic
whitespace (\s) with horizontal whitespace (\h).  This works because the
records in these subtitle files are delimited by vertical whitespace.